# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Expense:
    properties:
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: bhavani@example.com
    name: Bhavani
  description: RESTful API for tracking expenses
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
  /api/v1/expenses:
    get:
      description: Lists all expenses with optional filters
      parameters:
      - description: Category
        in: query
        name: category
        type: string
      - description: Currency
        in: query
        name: currency
        type: string
      - description: Start date (RFC3339)
        in: query
        name: from
        type: string
      - description: End date (RFC3339)
        in: query
        name: to
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Expense'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Adds a new expense record for the authenticated user
      parameters:
      - description: Expense object
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.Expense'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new expense
      tags:
      - expenses
  /api/v1/expenses/{id}:
    delete:
      description: Deletes an expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an expense
      tags:
      - expenses
    get:
      description: Returns a single expense by its UUID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an expense by ID
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Updates an existing expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated expense object
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.Expense'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an expense
      tags:
      - expenses
  /api/v1/expenses/summary:
    get:
      description: Provides category-wise expense summary for the user
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: from
        type: string
      - description: End date (RFC3339)
        in: query
        name: to
        type: string
      - description: Target currency (e.g. USD)
        in: query
        name: target_currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: number
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Expense summary
      tags:
      - expenses
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"